facial_clustering_app/requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dlib @ https://github.com/Arun-Raghav-S/face-clustering/raw/master/dlib-19.24.1-cp311-cp311-win_amd64.whl
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imutils==0.5.4
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
pydeck==0.9.0b1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0